requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
pyyaml==6.0.2
matplotlib==3.9.1
seaborn==0.13.2
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.5
pytest==8.3.2
Flask==3.0.3
